+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 CHANGELOG
 =========
 
+## 3.7.0 (2021-07-13)
+
+### Improvements
+
+- [sdk/nodejs] - Support for calling resource methods
+  [#7377](https://github.com/pulumi/pulumi/pull/7377)
+
+- [sdk/go] - Support for calling resource methods
+  [#7437](https://github.com/pulumi/pulumi/pull/7437)
+
+### Bug Fixes
+
+- [codegen/go] - Reimplement strict go enums to be Inputs.
+  [#7383](https://github.com/pulumi/pulumi/pull/7383)
+
+- [codegen/go] - Emit To[ElementType]Output methods for go enum output types
+  [#7499](https://github.com/pulumi/pulumi/pull/7499)
+
 ## 3.6.1 (2021-07-07)
 
 ### Improvements